requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
plotly>=5.0
Pillow>=10.0
requests>=2.31
scikit-learn>=1.3
SpeechRecognition>=3.10
toml>=0.10
